Drop packages: we can package version info now we get all info from the NuSpec files
