+++ src/DotNetProjectFile.Analyzers/Analyzers/MsBuild/UseAnalyzersForPackages.cs
@@ -36,13 +36,52 @@ public sealed class UseAnalyzersForPackages() : MsBuildProjectFileAnalyzer(Rule.
         }
     }
 
-    private static NuGet.Analyzer[] GetAnalyzers(string language)
+    private static Analyzer[] GetAnalyzers(string language)
         => Analyzers
         .Where(analyzer => analyzer.IsApplicable(language))
         .ToArray();
 
-    private static readonly NuGet.Analyzer[] Analyzers = NuGet.Packages.All
-        .OfType<NuGet.Analyzer>()
-        .Where(a => a.Match != "*")
-        .ToArray();
+    private static readonly Analyzer[] Analyzers =
+    [
+        new Analyzer("AwesomeAssertions.Analyzers", "AwesomeAssertions"),
+        new Analyzer("Ardalis.ApiEndpoints.CodeAnalyzers", "Ardalis.ApiEndpoints"),
+        new Analyzer("FakeItEasy.Analyzer.CSharp", "FakeItEasy", LanguageNames.CSharp),
+        new Analyzer("FakeItEasy.Analyzer.VisualBasic", "FakeItEasy", LanguageNames.VisualBasic),
+        new Analyzer("FluentAssertions.Analyzers", "FluentAssertions"),
+        new Analyzer("Libplanet.Analyzers", "Libplanet"),
+        new Analyzer("Lucene.Net.Analysis.Common", "Lucene.Net"),
+        new Analyzer("MassTransit.Analyzers", "MassTransit"),
+        new Analyzer("MessagePackAnalyzer", "MessagePack"),
+        new Analyzer("MessagePipe.Analyzer", "MessagePipe"),
+        new Analyzer("Microsoft.AspNetCore.Components.Analyzers", "Microsoft.AspNetCore.Components"),
+        new Analyzer("Microsoft.Azure.Functions.Analyzers", "Microsoft.Azure.Functions"),
+        new Analyzer("Microsoft.CodeAnalysis.Analyzers", "Microsoft.CodeAnalysis"),
+        new Analyzer("Microsoft.EntityFrameworkCore.Analyzers", "Microsoft.EntityFrameworkCore"),
+        new Analyzer("Microsoft.ServiceHub.Analyzers", "Microsoft.ServiceHub"),
+        new Analyzer("MongoDB.Analyzer", "MongoDB"),
+        new Analyzer("Moq.Analyzers", "Moq"),
+        new Analyzer("NSubstitute.Analyzers.CSharp", "NSubstitute", LanguageNames.CSharp),
+        new Analyzer("NSubstitute.Analyzers.VisualBasic", "NSubstitute", LanguageNames.VisualBasic),
+        new Analyzer("NUnit.Analyzers", "NUnit"),
+        new Analyzer("RuntimeContracts.Analyzer", "RuntimeContracts"),
+        new Analyzer("SerilogAnalyzer", "Serilog"),
+        new Analyzer("xunit.analyzers", "xunit"),
+        new Analyzer("ZeroFormatter.Analyzer", "ZeroFormatter"),
+    ];
+
+    private sealed record Analyzer(string Name, string Match, string? Language = null)
+    {
+        public bool IsApplicable(string compilationLanguage)
+            => Language is null || Language == compilationLanguage;
+
+        public bool IsAnalyzerFor(CachedPackage pkg)
+            => IsAnalyzerFor(pkg.Name);
+        
+        public bool IsAnalyzerFor(string name)
+            => name.StartsWith(Match, StringComparison.OrdinalIgnoreCase)
+            && (name.Length == Match.Length
+                || name[Match.Length] == '.');
+
+        public bool IsMatch(PackageReferenceBase reference) => reference.Include.IsMatch(Name);
+    }
 }
