--- src/DotNetProjectFile.Analyzers/NuGet/BuildExtension.cs
@@ -1,7 +1,0 @@
-namespace DotNetProjectFile.NuGet;
-
-public sealed record BuildExtension : Package
-{
-    public BuildExtension(string name, string? language = null)
-        : base(name, language, isPrivateAsset: true) { }
-}
